refactor: split ChatBubble.swift into type-specific content views
