+++ clients/macos/vellum-assistant/Features/Chat/ChatBubbleAttachmentContent.swift
@@ -1,0 +1,143 @@
+import SwiftUI
+import VellumAssistantShared
+
+// MARK: - Attachment Content
+
+extension ChatBubble {
+    var attachmentSummary: String {
+        let count = message.attachments.count
+        if count == 1 {
+            return "Sent \(message.attachments[0].filename)"
+        }
+        return "Sent \(count) attachments"
+    }
+
+    /// Partitions attachments into decoded images, videos, and non-media files in a single pass,
+    /// avoiding redundant base64 decoding and NSImage construction across render calls.
+    var partitionedAttachments: (images: [(ChatAttachment, NSImage)], videos: [ChatAttachment], files: [ChatAttachment]) {
+        var images: [(ChatAttachment, NSImage)] = []
+        var videos: [ChatAttachment] = []
+        var files: [ChatAttachment] = []
+        for attachment in message.attachments {
+            if attachment.mimeType.hasPrefix("image/"), let img = nsImage(for: attachment) {
+                images.append((attachment, img))
+            } else if attachment.mimeType.hasPrefix("video/") {
+                videos.append(attachment)
+            } else {
+                files.append(attachment)
+            }
+        }
+        return (images, videos, files)
+    }
+
+    func attachmentImageGrid(_ images: [(ChatAttachment, NSImage)]) -> some View {
+        let columns = images.count == 1
+            ? [GridItem(.flexible())]
+            : [GridItem(.flexible(), spacing: VSpacing.sm), GridItem(.flexible(), spacing: VSpacing.sm)]
+        return LazyVGrid(columns: columns, alignment: .leading, spacing: VSpacing.sm) {
+            ForEach(images, id: \.0.id) { attachment, nsImage in
+                Image(nsImage: nsImage)
+                    .resizable()
+                    .aspectRatio(contentMode: images.count == 1 ? .fit : .fill)
+                    .frame(maxHeight: images.count == 1 ? 320 : 180)
+                    .clipped()
+                    .clipShape(RoundedRectangle(cornerRadius: VRadius.sm))
+                    .onTapGesture {
+                        openImageInPreview(attachment)
+                    }
+            }
+        }
+    }
+
+    func fileAttachmentChip(_ attachment: ChatAttachment) -> some View {
+        HStack(spacing: VSpacing.xs) {
+            Image(systemName: fileIcon(for: attachment.mimeType))
+                .font(VFont.caption)
+                .foregroundColor(isUser ? VColor.userBubbleTextSecondary : VColor.textSecondary)
+
+            Text(attachment.filename)
+                .font(VFont.caption)
+                .foregroundColor(isUser ? VColor.userBubbleText : VColor.textPrimary)
+                .lineLimit(1)
+
+            Text(formattedFileSize(base64Length: attachment.dataLength))
+                .font(VFont.small)
+                .foregroundColor(isUser ? VColor.userBubbleTextSecondary : VColor.textMuted)
+        }
+        .padding(.horizontal, VSpacing.sm)
+        .padding(.vertical, VSpacing.xs)
+        .background(
+            RoundedRectangle(cornerRadius: VRadius.sm)
+                .fill(isUser ? VColor.userBubbleText.opacity(0.15) : VColor.surfaceBorder.opacity(0.5))
+        )
+        .contentShape(Rectangle())
+        .onTapGesture {
+            saveFileAttachment(attachment)
+        }
+        .onHover { hovering in
+            if hovering {
+                NSCursor.pointingHand.push()
+            } else {
+                NSCursor.pop()
+            }
+        }
+    }
+
+    func nsImage(for attachment: ChatAttachment) -> NSImage? {
+        // Use pre-decoded thumbnail image — avoids NSImage(data:) during layout, which
+        // can trigger re-entrant AppKit constraint invalidation and crash on scroll.
+        if let img = attachment.thumbnailImage {
+            return img
+        }
+        if let thumbnailData = attachment.thumbnailData, let img = NSImage(data: thumbnailData) {
+            return img
+        }
+        if let data = Data(base64Encoded: attachment.data), let img = NSImage(data: data) {
+            return img
+        }
+        return nil
+    }
+
+    func openImageInPreview(_ attachment: ChatAttachment) {
+        guard let data = Data(base64Encoded: attachment.data) else { return }
+        let tempDir = FileManager.default.temporaryDirectory
+        let sanitized = (attachment.filename as NSString).lastPathComponent
+        let fileURL = tempDir.appendingPathComponent(sanitized.isEmpty ? "image" : sanitized)
+        do {
+            try data.write(to: fileURL)
+            NSWorkspace.shared.open(fileURL)
+        } catch {
+            // Silently fail — not critical
+        }
+    }
+
+    func saveFileAttachment(_ attachment: ChatAttachment) {
+        guard let data = Data(base64Encoded: attachment.data) else { return }
+        let panel = NSSavePanel()
+        panel.nameFieldStringValue = (attachment.filename as NSString).lastPathComponent
+        panel.canCreateDirectories = true
+        panel.begin { response in
+            guard response == .OK, let url = panel.url else { return }
+            try? data.write(to: url)
+        }
+    }
+
+    func fileIcon(for mimeType: String) -> String {
+        if mimeType.hasPrefix("video/") { return "film" }
+        if mimeType.hasPrefix("audio/") { return "waveform" }
+        if mimeType.hasPrefix("text/") { return "doc.text.fill" }
+        if mimeType == "application/pdf" { return "doc.fill" }
+        if mimeType.contains("zip") || mimeType.contains("archive") { return "doc.zipper" }
+        if mimeType.contains("json") || mimeType.contains("xml") { return "doc.text.fill" }
+        return "doc.fill"
+    }
+
+    func formattedFileSize(base64Length: Int) -> String {
+        let bytes = base64Length * 3 / 4
+        if bytes < 1024 { return "\(bytes) B" }
+        let kb = Double(bytes) / 1024
+        if kb < 1024 { return String(format: "%.1f KB", kb) }
+        let mb = kb / 1024
+        return String(format: "%.1f MB", mb)
+    }
+}
