+++ clients/macos/vellum-assistant/Features/Chat/ChatBubbleInterleavedContent.swift
@@ -1,0 +1,94 @@
+import SwiftUI
+import VellumAssistantShared
+
+// MARK: - Interleaved Content
+
+extension ChatBubble {
+    /// Whether this message has meaningful interleaved content (multiple block types).
+    var hasInterleavedContent: Bool {
+        // Use interleaved path when contentOrder has more than one distinct block type
+        guard message.contentOrder.count > 1 else { return false }
+        var hasTextBlock = false
+        var hasNonText = false
+        for ref in message.contentOrder {
+            switch ref {
+            case .text: hasTextBlock = true
+            case .toolCall, .surface: hasNonText = true
+            }
+            if hasTextBlock && hasNonText { return true }
+        }
+        return false
+    }
+
+    /// Groups consecutive tool call refs for rendering.
+    enum ContentGroup {
+        case text(Int)
+        case toolCalls([Int])
+        case surface(Int)
+    }
+
+    func groupContentBlocks() -> [ContentGroup] {
+        var groups: [ContentGroup] = []
+        for ref in message.contentOrder {
+            switch ref {
+            case .text(let i):
+                groups.append(.text(i))
+            case .toolCall(let i):
+                if case .toolCalls(let indices) = groups.last {
+                    groups[groups.count - 1] = .toolCalls(indices + [i])
+                } else {
+                    groups.append(.toolCalls([i]))
+                }
+            case .surface(let i):
+                groups.append(.surface(i))
+            }
+        }
+        return groups
+    }
+
+    @ViewBuilder
+    var interleavedContent: some View {
+        let groups = groupContentBlocks()
+
+        // Render all content groups in order: text, tool calls, and surfaces
+        ForEach(Array(groups.enumerated()), id: \.offset) { _, group in
+            switch group {
+            case .text(let i):
+                if i < message.textSegments.count {
+                    let segmentText = message.textSegments[i].trimmingCharacters(in: .whitespacesAndNewlines)
+                    if !segmentText.isEmpty {
+                        textBubble(for: segmentText)
+                    }
+                }
+            case .toolCalls:
+                // Tool calls are rendered by trailingStatus below the message
+                EmptyView()
+            case .surface(let i):
+                if i < message.inlineSurfaces.count,
+                   message.inlineSurfaces[i].id != taskProgressOverlay.activeSurfaceId {
+                    InlineSurfaceRouter(surface: message.inlineSurfaces[i], onAction: onSurfaceAction)
+                }
+            }
+        }
+
+        // Attachments are not part of contentOrder but must still be rendered
+        let partitioned = partitionedAttachments
+        if !partitioned.images.isEmpty {
+            attachmentImageGrid(partitioned.images)
+        }
+        if !partitioned.videos.isEmpty {
+            VStack(alignment: .leading, spacing: VSpacing.sm) {
+                ForEach(partitioned.videos) { attachment in
+                    InlineVideoAttachmentView(attachment: attachment, daemonHttpPort: daemonHttpPort)
+                }
+            }
+        }
+        if !partitioned.files.isEmpty {
+            VStack(alignment: .leading, spacing: VSpacing.xs) {
+                ForEach(partitioned.files) { attachment in
+                    fileAttachmentChip(attachment)
+                }
+            }
+        }
+    }
+}
